refactor: project renamed to funny cats
