+++ app/src/androidTest/java/com/hfad/android/funnycats/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.hfad.android.hotcats
+package com.hfad.android.funnycats
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
