+++ app/src/main/java/com/hfad/android/funnycats/model/Cat.kt
@@ -1,4 +1,4 @@
-package com.hfad.android.hotcats.model
+package com.hfad.android.funnycats.model
 
 
 data class Cat(val imageUrl:String?,val title:String?){
