+++ app/src/main/java/com/hfad/android/funnycats/viemodels/CatListViewModel.kt
@@ -1,4 +1,4 @@
-package com.hfad.android.hotcats.viemodels
+package com.hfad.android.funnycats.viemodels
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
@@ -6,9 +6,9 @@ import androidx.lifecycle.viewModelScope
 import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.paging.PagingData
-import com.hfad.android.hotcats.data.CatApi
-import com.hfad.android.hotcats.model.Cat
-import com.hfad.android.hotcats.paging.CatPageSource
+import com.hfad.android.funnycats.data.CatApi
+import com.hfad.android.funnycats.model.Cat
+import com.hfad.android.funnycats.paging.CatPageSource
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.stateIn
