+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Hot Cats</string>
+    <string name="app_name">Funny Cats</string>
 </resources>
