+++ app/src/test/java/com/hfad/android/funnycats/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.hfad.android.hotcats
+package com.hfad.android.funnycats
 
 import org.junit.Test
